+++ core/designsystem/src/main/java/com/eshc/goonersapp/core/designsystem/component/MatchLeagueInfo.kt
@@ -61,7 +61,7 @@ fun MatchLeagueInfo(
         Text(
             text = competitionName,
             color = ColorFF181818,
-            style = GnrTypography.descriptionSemiBold
+            style = GnrTypography.descriptionMedium
         )
     }
 }
